requirements.txt
asgiref==3.4.1
Babel==2.9.1
dj-database-url==0.5.0
Django==3.2.10
django-heroku==0.3.1
django-money==2.1
djangorestframework==3.13.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.1.0
psycopg2==2.9.2
py-moneyed==1.2
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.3.0
